New RemovedHexadecimalNumericStrings sniff, split off from ValidIntegers
Split the `ValidIntegers` sniff in two:
* A new `PHPCompatibility.Numbers.RemovedHexadecimalNumericStrings` which examines hexadecimal numeric strings.
* The existing `ValidIntegers` sniff will now only focus on actual numbers (binary, octal).

This is a BC-break and should be annotated as such in the changelog.

Note: the new sniff implements the PHPCSUtils `Numbers::isHexidecimalInt()` method.
