+++ PHPCompatibility/Tests/Numbers/RemovedHexadecimalNumericStringsUnitTest.inc
@@ -1,0 +1,6 @@
+<?php
+
+$hex = 'aa78b5'; // Ok.
+
+$hex = '0xaa78b5';
+$hexUpper = '0Xbb99EF';
