+++ PHPCompatibility/Tests/Numbers/ValidIntegersUnitTest.inc
@@ -8,8 +8,4 @@ $invalidOctal = 08;
 $invalidOctal = 038;
 $invalidOctal = 091;
 
-$hex = '0xaa78b5';
-$hex = 'aa78b5'; // Ok.
-
 $binaryUpper = 0B10001;
-$hexUpper = '0Xbb99EF';
