+++ PHPCompatibility/Tests/Numbers/ValidIntegersUnitTest.php
@@ -61,7 +61,7 @@ class ValidIntegersUnitTest extends BaseSniffTest
         return array(
             array(3, '0b001001101', true),
             array(4, '0b01', false),
-            array(14, '0B10001', true),
+            array(11, '0B10001', true),
         );
     }
 
@@ -125,62 +125,6 @@ class ValidIntegersUnitTest extends BaseSniffTest
     {
         $file = $this->sniffFile(__FILE__, '4.0-99.0');
         $this->assertNoViolation($file, 6);
-    }
-
-
-    /**
-     * testHexNumericString
-     *
-     * @dataProvider dataHexNumericString
-     *
-     * @param int    $line Line number where the error should occur.
-     * @param string $hex  Hexidecminal number as a string.
-     *
-     * @return void
-     */
-    public function testHexNumericString($line, $hex)
-    {
-        $error = "The behaviour of hexadecimal numeric strings was inconsistent prior to PHP 7 and support has been removed in PHP 7. Found: '{$hex}'";
-
-        $file = $this->sniffFile(__FILE__, '5.6');
-        $this->assertWarning($file, $line, $error);
-
-        $file = $this->sniffFile(__FILE__, '7.0');
-        $this->assertError($file, $line, $error);
-    }
-
-    /**
-     * Data provider.
-     *
-     * @see testHexNumericString()
-     *
-     * @return array
-     */
-    public function dataHexNumericString()
-    {
-        // phpcs:disable PHPCompatibility.Numbers.ValidIntegers.HexNumericStringFound
-        return array(
-            array(11, '0xaa78b5'),
-            array(15, '0Xbb99EF'),
-        );
-        // phpcs:enable
-    }
-
-
-    /**
-     * testHexNumericString.
-     *
-     * @dataProvider dataHexNumericString
-     *
-     * @return void
-     */
-    public function testNoFalsePositivesHexNumericString()
-    {
-        $file = $this->sniffFile(__FILE__, '5.6');
-        $this->assertNoViolation($file, 12);
-
-        $file = $this->sniffFile(__FILE__, '7.0');
-        $this->assertNoViolation($file, 12);
     }
